# backend/src/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "S.O.S Cidade API",
    "description": "API do sistema S.O.S Cidade para gestão de denúncias e ocorrências.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Local Development Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "token"
      }
    },
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "email": { "type": "string" },
          "name": { "type": "string" },
          "role": { "type": "string", "enum": ["DENUNCIANTE", "GESTOR"] },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
      "Denuncia": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "tipoDenuncia": { "type": "string" },
          "identificacao": { "type": "boolean" },
          "nomeDenunciante": { "type": "string" },
          "userId": { "type": "integer", "nullable": true },
          "bairroOcorrencia": { "type": "string" },
          "descricaoOcorrencia": { "type": "string" },
          "dataOcorrencia": { "type": "string" },
          "protocolo": { "type": "string" },
          "situacao": { "type": "string" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    },
    {
      "cookieAuth": []
    }
  ],
  "paths": {
    "/auth/login": {
      "post": {
        "summary": "Autenticar usuário",
        "tags": ["Auth"],
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": { "type": "string" },
                  "password": { "type": "string" }
                },
                "required": ["email", "password"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Usuário autenticado com sucesso e cookie setado",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "user": { "$ref": "#/components/schemas/User" },
                    "token": { "type": "string" }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Credenciais inválidas"
          }
        }
      }
    },
    "/auth/logout": {
      "post": {
        "summary": "Sair da aplicação (limpar cookie)",
        "tags": ["Auth"],
        "security": [],
        "responses": {
          "200": {
            "description": "Logout realizado com sucesso"
          }
        }
      }
    },
    "/users": {
      "post": {
        "summary": "Criar um novo usuário",
        "tags": ["Users"],
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": { "type": "string" },
                  "name": { "type": "string" },
                  "password": { "type": "string" },
                  "role": { "type": "string" }
                },
                "required": ["email", "name", "password"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Usuário criado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          }
        }
      }
    },
    "/denuncias": {
      "get": {
        "summary": "Listar denúncias com paginação",
        "tags": ["Denuncias"],
        "parameters": [
          {
            "name": "_page",
            "in": "query",
            "schema": { "type": "integer", "default": 1 },
            "description": "Número da página"
          },
          {
            "name": "_limit",
            "in": "query",
            "schema": { "type": "integer", "default": 50 },
            "description": "Quantidade de itens por página"
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de denúncias e metadados de paginação",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/Denuncia" }
                    },
                    "meta": {
                      "type": "object",
                      "properties": {
                        "total": { "type": "integer" },
                        "page": { "type": "integer" },
                        "limit": { "type": "integer" }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Criar uma nova denúncia",
        "tags": ["Denuncias"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "tipoDenuncia": { "type": "string" },
                  "identificacao": { "type": "boolean" },
                  "bairroOcorrencia": { "type": "string" },
                  "descricaoOcorrencia": { "type": "string" },
                  "dataOcorrencia": { "type": "string" }
                },
                "required": ["tipoDenuncia", "identificacao", "bairroOcorrencia", "descricaoOcorrencia", "dataOcorrencia"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Denúncia criada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Denuncia"
                }
              }
            }
          }
        }
      }
    },
    "/denuncias/{id}": {
      "get": {
        "summary": "Buscar denúncia por ID",
        "tags": ["Denuncias"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "200": {
            "description": "Denúncia encontrada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Denuncia"
                }
              }
            }
          },
          "404": {
            "description": "Denúncia não encontrada"
          }
        }
      },
      "patch": {
        "summary": "Atualizar uma denúncia",
        "tags": ["Denuncias"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "situacao": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Denúncia atualizada"
          }
        }
      },
      "delete": {
        "summary": "Deletar uma denúncia",
        "tags": ["Denuncias"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "200": {
            "description": "Denúncia deletada"
          }
        }
      }
    }
  }
}